@@ -17,11 +17,11 @@
     <description>Spring PetClinic Vets Service</description>
 
     <properties>
-		<java.version>17</java.version>
+		<java.version>21</java.version>
 		<spring-cloud.version>2024.0.0</spring-cloud.version>
-		<chaos-monkey-spring-boot.version>3.1.0</chaos-monkey-spring-boot.version>
-		<jolokia-core.version>1.7.1</jolokia-core.version>
-		<version.spring.cloud.azure>5.20.1</version.spring.cloud.azure>
+		<chaos-monkey-spring-boot.version>4.0.0</chaos-monkey-spring-boot.version>
+		<jolokia-core.version>2.5.0</jolokia-core.version>
+		<version.spring.cloud.azure>7.0.0</version.spring.cloud.azure>
         <docker.image.exposed.port>8081</docker.image.exposed.port>
         <docker.image.dockerfile.dir>${basedir}/../docker</docker.image.dockerfile.dir>
 		<docker.plugin.version>1.2.0</docker.plugin.version>
